JavaDoc + removed warnings
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@25199 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/kernel/commands/CmdDensityPlot.java
@@ -12,9 +12,9 @@ import geogebra.common.main.MyError;
 /**
  * Giuliano Bellucci 05/04/2013
  * 
- * densityplot[2-variables function]
- * densityplot[2-variables function,minx,maxx,miny,maxy] 
- * densityplot[2-variables function,minx,maxx,miny,maxy,quality] quality=2 for web 1 for desktop
+ * densityplot[2-variables function] densityplot[2-variables
+ * function,minx,maxx,miny,maxy] densityplot[2-variables
+ * function,minx,maxx,miny,maxy,quality] quality=2 for web 1 for desktop
  * 
  */
 
@@ -26,6 +26,12 @@ public class CmdDensityPlot extends CommandProcessor {
 	private double highY;
 	private GeoElement[] args;
 
+	/**
+	 * Creates new command processor
+	 * 
+	 * @param kernel
+	 *            kernel
+	 */
 	public CmdDensityPlot(Kernel kernel) {
 		super(kernel);
 	}
@@ -33,7 +39,7 @@ public class CmdDensityPlot extends CommandProcessor {
 	@Override
 	public GeoElement[] process(Command c) throws MyError,
 			CircularDefinitionException {
-		
+
 		int n = c.getArgumentNumber();
 
 		switch (n) {
@@ -59,8 +65,10 @@ public class CmdDensityPlot extends CommandProcessor {
 			return ret;
 		case 6:
 			control(c);
-			// Quality = 1 -> Image size = 280 x 280. 1 pixel for 1 function value
-			// Quality = 2 -> Image size = 140 x 140. 4 pixels for 1 function value (for web)
+			// Quality = 1 -> Image size = 280 x 280. 1 pixel for 1 function
+			// value
+			// Quality = 2 -> Image size = 140 x 140. 4 pixels for 1 function
+			// value (for web)
 			int quality = (int) c.getArgument(5).evaluateNum().getNumber()
 					.getDouble();
 			if (Double.isNaN(quality) || (quality != 1 && quality != 2)) {
@@ -79,28 +87,32 @@ public class CmdDensityPlot extends CommandProcessor {
 			if (Double.isNaN(quality) || (quality != 1 && quality != 2)) {
 				throw argErr(app, c.getName(), c.getArgument(5));
 			}
-			double zeroX=(double) c.getArgument(6).evaluateNum().getNumber().getDouble();
+			double zeroX = c.getArgument(6).evaluateNum().getNumber()
+					.getDouble();
 			if (Double.isNaN(zeroX)) {
 				throw argErr(app, c.getName(), c.getArgument(6));
 			}
-			double zeroY=(double) c.getArgument(7).evaluateNum().getNumber().getDouble();
+			double zeroY = c.getArgument(7).evaluateNum().getNumber()
+					.getDouble();
 			if (Double.isNaN(zeroY)) {
 				throw argErr(app, c.getName(), c.getArgument(7));
 			}
-			boolean hasAxes=false;
-			if (c.getArgument(8).isBooleanValue()){
-				hasAxes="true".equals(c.getArgument(8).toValueString(StringTemplate.defaultTemplate));
-			}else{
+			boolean hasAxes = false;
+			if (c.getArgument(8).isBooleanValue()) {
+				hasAxes = "true".equals(c.getArgument(8).toValueString(
+						StringTemplate.defaultTemplate));
+			} else {
 				throw argErr(app, c.getName(), c.getArgument(8));
 			}
-			boolean hasGrid=false;
-			if (c.getArgument(9).isBooleanValue()){
-				hasGrid="true".equals(c.getArgument(9).toValueString(StringTemplate.defaultTemplate));
-			}else{
+			boolean hasGrid = false;
+			if (c.getArgument(9).isBooleanValue()) {
+				hasGrid = "true".equals(c.getArgument(9).toValueString(
+						StringTemplate.defaultTemplate));
+			} else {
 				throw argErr(app, c.getName(), c.getArgument(9));
 			}
 			algo = new AlgoDensityPlot(cons, (GeoFunctionNVar) args[0], lowX,
-					highX, lowY, highY, quality,zeroX,zeroY,hasAxes,hasGrid);
+					highX, lowY, highY, quality, zeroX, zeroY, hasAxes, hasGrid);
 			algo.compute();
 			ret = new GeoElement[1];
 			ret[0] = algo.getResult();
@@ -127,5 +139,5 @@ public class CmdDensityPlot extends CommandProcessor {
 			throw argErr(app, c.getName(), c.getArgument(3));
 		}
 	}
-	
+
 }
